Cleans up lots of miscellaneous code.
