+++ edit-client/src/client.rs
@@ -4,7 +4,6 @@ mod state;
 pub use self::actions::*;
 pub use self::state::*;
 
-use oatie;
 use crate::{
     random::*,
     walkers::Pos,
@@ -15,6 +14,7 @@ use edit_common::{
     markdown::doc_to_markdown,
 };
 use failure::Error;
+use oatie;
 use oatie::{
     doc::*,
     validate::validate_doc,
@@ -258,7 +258,10 @@ pub fn button_handlers<C: ClientController>(
     (callbacks, ui)
 }
 
-fn controller_command<C: ClientController>(client: &mut C, req: ControllerCommand) -> Result<(), Error> {
+fn controller_command<C: ClientController>(
+    client: &mut C,
+    req: ControllerCommand,
+) -> Result<(), Error> {
     match req {
         ControllerCommand::RenameGroup { tag, curspan: _ } => {
             client.client_op(|doc| replace_block(doc, &tag))?;
@@ -474,8 +477,7 @@ pub trait ClientController {
                                 .state()
                                 .client_doc
                                 .sync_confirmed_pending_op(&doc, version);
-                            if let Some(local_op) = local_op
-                            {
+                            if let Some(local_op) = local_op {
                                 // Send our next operation.
                                 self.upload(local_op)?;
                             }
@@ -492,7 +494,9 @@ pub trait ClientController {
                             // Client drives frontend frontend state.
                             let res = if cfg!(feature = "DEBUG_full_client_updates") {
                                 // Fully refresh the client.
-                                FrontendCommand::RenderFull(doc_as_html(&self.state().client_doc.doc.0))
+                                FrontendCommand::RenderFull(doc_as_html(
+                                    &self.state().client_doc.doc.0,
+                                ))
                             } else {
                                 // Render delta.
                                 FrontendCommand::RenderDelta(
@@ -524,7 +528,8 @@ pub trait ClientController {
 
                     Task::ClientCommand(ClientCommand::ServerDisconnect) => {
                         // Notify frontend.
-                        self.send_frontend(&FrontendCommand::ServerDisconnect).unwrap();
+                        self.send_frontend(&FrontendCommand::ServerDisconnect)
+                            .unwrap();
                     }
                 }
 
@@ -562,10 +567,7 @@ pub trait ClientController {
         let doc = self.state().client_doc.doc.clone();
         let client_id = self.state().client_id.clone();
 
-        callback(ActionContext {
-            doc,
-            client_id,
-        })
+        callback(ActionContext { doc, client_id })
     }
 
     fn client_op<C>(&mut self, callback: C) -> Result<(), Error>
